Partially revert commit 2b8ae34 to keep documentation
Having this feature documented in the mariadb postinst and
postrm scripts is needed so that the same packaging code
is easier to understand and re-use in downstream and upstream
packaging where the new mysql-common package might not
be available.

For the same reason, echo the mv command tips for users if
the packaging is used on a platform that does not have
the new mysql-common package.

Checking for the existance of the configure-symlink hook
on postrm is also useful for scenarios where the mysql-common
package gets purged before the mariadb-server-10.0 is purged.