Remove all old passwordless root account lines when using unix auth
This closes a potential security vulnerability. In bug report at
https://bugs.launchpad.net/ubuntu/+source/mariadb-10.0/+bug/1561062
it was pointed out, that is name resolve is skipped and TCP is
forced, localhost users can being detected as connecting via host
localhost and the passwordless root lines would apply and let
user inside. This is not intended and potentially dangerous,
so remove the lines. They are most likely useless and legacy
anyway.