Remove mariadb-plugin-cassandra
The Cassandra plugin requires libthrift-dev, which isn't available in
Debian and thus the plugin hasn't really been built in Debian for
a while. Clean it away to keep the control file sleek.

This also helps on future uploads as the mariadb-plugin-cassandra
mentioned in the control file will not trigger NEW queue processing.