Re-add libmariadbclient18 and libmariadbclient-dev
Add libmariadbclient.so.18 and related files. This is MariaDB's fork of
MySQL's libmysqlclient.so.18. We rename the soname to
libmariadbclient.so.18 to avoid a conflict with a similar file produced
by MySQL, as MySQL upstream also continues the line with
libmysqlclient.so.20.

This allows packages to build against libmariadbclient-dev (using
-lmariadbclient) if they want to avoid a dependency on a package built
from MySQL sources.

We conflict with libmysqlclient-dev since we provide the same header
files but from MariaDB's fork.

This is a partial reversion of commit
24ed76b33cb076ef886b9637d96aa310cb8efa88.