Make libmariadbclient18 depend on mysql-common only (Closes: #850216)
Both libmariadbclientXX and libmysqlclientXX should depend on mysql-common
alone, and thus they will end up being installed with the fallback
configuration in the absence of other client or server packages that
would pull in some of the MariaDB or Oracle MySQL specific configs.

If libmariadbclientXX depends on mariadb-common, that would pull in
config files that can create the situation described in Bug#850216.
Instead if should depend on mysql-client directly and only on it.