Fix for #848616, mysql_install_db creates unusable root user
An earlier change implemented that mariadb-server-10.1 installs the
default root user with unix socket authentication. However, this
change was done in a way that breaks /usr/bin/mysql_install_db for
other users.

This commit adds a patch that instead extends mysql_install_db with
options to configure how the initial root account is created. This
allows mariadb-server-10.1.postinst to use those options to get the
new behaviour, while preserving backwards compatibility with other
uses.

The patch is destined for upstream MariaDB 10.1.21, so should be
possible to remove again from debian/patches/ shortly.

This commit also cleans up in a similar way the removal of the
anonymous user, which was using a similar mechanism.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>